+++ src/UIKit/UITypes.cs
@@ -102,83 +102,6 @@ namespace UIKit {
 #endif
 	}
 
-	[Watch (4,0), TV (11,0), iOS (11,0)]
-	[StructLayout (LayoutKind.Sequential)]
-	public struct NSDirectionalEdgeInsets {
-
-		// API match for NSDirectionalEdgeInsetsZero field/constant
-		[Field ("NSDirectionalEdgeInsetsZero")] // fake (but helps testing and could also help documentation)
-		public static readonly NSDirectionalEdgeInsets Zero;
-
-		public nfloat Top, Leading, Bottom, Trailing;
-
-#if !COREBUILD
-		public NSDirectionalEdgeInsets (nfloat top, nfloat leading, nfloat bottom, nfloat trailing)
-		{
-			Top = top;
-			Leading = leading;
-			Bottom = bottom;
-			Trailing = trailing;
-		}
-
-		// note: NSDirectionalEdgeInsetsEqualToDirectionalEdgeInsets (UIGeometry.h) is a macro
-		public bool Equals (NSDirectionalEdgeInsets other)
-		{
-			if (Leading != other.Leading)
-				return false;
-			if (Trailing != other.Trailing)
-				return false;
-			if (Top != other.Top)
-				return false;
-			return (Bottom == other.Bottom);
-		}
-
-		public override bool Equals (object obj)
-		{
-			if (obj is NSDirectionalEdgeInsets)
-				return Equals ((NSDirectionalEdgeInsets) obj);
-			return false;
-		}
-
-		public static bool operator == (NSDirectionalEdgeInsets insets1, NSDirectionalEdgeInsets insets2)
-		{
-			return insets1.Equals (insets2);
-		}
-
-		public static bool operator != (NSDirectionalEdgeInsets insets1, NSDirectionalEdgeInsets insets2)
-		{
-			return !insets1.Equals (insets2);
-		}
-
-		public override int GetHashCode ()
-		{
-			return Top.GetHashCode () ^ Leading.GetHashCode () ^ Trailing.GetHashCode () ^ Bottom.GetHashCode ();
-		}
-
-		[DllImport (Constants.UIKitLibrary)]
-		extern static NSDirectionalEdgeInsets NSDirectionalEdgeInsetsFromString (IntPtr /* NSString */ s);
-
-		static public NSDirectionalEdgeInsets FromString (string s)
-		{
-			// note: null is allowed
-			var ptr = NSString.CreateNative (s);
-			var value = NSDirectionalEdgeInsetsFromString (ptr);
-			NSString.ReleaseNative (ptr);
-			return value;
-		}
-
-		[DllImport (Constants.UIKitLibrary)]
-		extern static IntPtr /* NSString */ NSStringFromDirectionalEdgeInsets (NSDirectionalEdgeInsets insets);
-
-		// note: ensure we can roundtrip ToString into FromString
-		public override string ToString ()
-		{
-			using (var ns = new NSString (NSStringFromDirectionalEdgeInsets (this)))
-				return ns.ToString ();
-		}
-#endif
-	}
-
 #if !WATCH
 	[iOS (9,0)]
 	[StructLayout (LayoutKind.Sequential)]
