+++ src/frameworks.sources
@@ -70,6 +70,7 @@ APPKIT_CORE_SOURCES = \
 	AppKit/Enums.cs \
 	AppKit/XEnums.cs \
 	AppKit/NSAccessibility.cs \
+	XKit/Types.cs \
 
 APPKIT_SOURCES = \
 	AppKit/ActionDispatcher.cs \
@@ -1499,6 +1500,7 @@ UIKIT_CORE_SOURCES = \
     UIKit/UIPrintInteractionController.cs \
     UIKit/UIStringAttributes.cs \
     UIKit/UITypes.cs \
+	XKit/Types.cs \
 
 UIKIT_SOURCES = \
 	UIKit/Compat.cs \
